Fixed some compiler warnings

+++ app/app1.cpp
@@ -19,14 +19,14 @@ int main() {
 
     std::string msg = "GET / HTTP/1.1\r\n\r\n";
 
-    if (!socket.send_data(msg.c_str(), msg.size())) {
+    if (!socket.send_data( msg.c_str(), static_cast<int>(msg.size()) )) {
         std::cout << "Failed to send data\n";
         return -1;
     }
 
     std::vector<char> server_reply(1024 * 8);
 
-    if (!socket.recieve_data(server_reply.data(), server_reply.size())) {
+    if (!socket.recieve_data( server_reply.data(), static_cast<int>(server_reply.size()) )) {
         std::cout << "Failed to recieve data\n";
         return -1;
     }
